+++ example/lib/main.dart
@@ -33,7 +33,7 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   final CurrencyTextFieldController _controller = CurrencyTextFieldController();
   final CurrencyTextFieldController _controller2 =
-      CurrencyTextFieldController(initDoubleValue: -10);
+      CurrencyTextFieldController(initDoubleValue: -10,currencySeparator: ' -> ');
   final CurrencyTextFieldController _controller3 =
       CurrencyTextFieldController(initIntValue: -1000,enableNegative: false);